+++ docs/CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 -----
 - Python execution engine
 - Support multiple entity types
+- Support extending viewsets with custom filter methods
 
 Fixed
 -----